do a lil refactor
